Widen ids to 32bits, move gasprice to offerDetail, narrow density to 112bits
This extends ids to 4billion instead of 16million, increasing the time
until they have been exhausted.

+++ contracts/MgvRoot.sol
@@ -53,7 +53,7 @@ contract MgvRoot is HasMgvEvents {
 
   /* Checking the size of `density` is necessary to prevent overflow when `density` is used in calculations. */
   function checkDensity(uint density) internal pure returns (bool) {
-    return uint128(density) == density;
+    return uint112(density) == density;
   }
 
   /* Checking the size of `gasprice` is necessary to prevent a) data loss when `gasprice` is copied to an `OfferDetail` struct, and b) overflow when `gasprice` is used in calculations. */
